+++ job_config/src/lib.rs
@@ -20,4 +20,5 @@
 // post = "cleanup.sh"
 
 pub mod config;
+pub mod job;
 pub mod workflow;